python/pyproject.toml
[tool.poetry]
name = "crypto-analytics-engine"
version = "0.1.0"
description = "Statistical arbitrage backtest engine"
authors = ["Luifer <your-email@example.com>"]
readme = "README.md"
package-mode = false  # We're not building a package, just running an app

[tool.poetry.dependencies]
python = "^3.11"
# Core numerical
numpy = "^1.26"
pandas = "^2.2"
scipy = "^1.12"

# Statistical analysis
statsmodels = "^0.14"
pykalman = "^0.9"

# API
fastapi = "^0.109"
uvicorn = {extras = ["standard"], version = "^0.27"}
pydantic = "^2.6"
python-multipart = "^0.0.9"

# Utils
python-dotenv = "^1.0"
loguru = "^0.7"

# Performance (now with LLVM installed)
numba = "^0.59"
vectorbt = "^0.26"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
httpx = "^0.27"  # For testing FastAPI
black = "^24.0"
ruff = "^0.2"
mypy = "^1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
